backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
opencv-python-headless
requests
python-dotenv

gunicorn
PyJWT